requirements-dev.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.1.0
python-dotenv>=1.0.0

# MLOps
mlflow>=2.0.0
dvc>=2.0.0
wandb>=0.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0

# Code Quality
ruff==0.1.9
mypy==1.8.0

# Security
bandit==1.7.6

# Pre-commit
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings[python]==0.24.0

# Development utilities
ipython==8.18.1
ipdb==0.13.13

# Logging
loguru>=0.6.0

# kaggle
kagglehub

fastapi
uvicorn
